Move BootOption handling to ComputerSystem
BootOptions are not a top level object, and only used under the context
of a ComputerSystem. This moves its definition to be with the
ComputerSystem and related structs, and removes unnecessary fields and
json unmarshal handling.
