+++ redfish/computersystem.go
@@ -345,6 +345,51 @@ func (boot *Boot) UnmarshalJSON(b []byte) error {
 	return nil
 }
 
+// BootOption represents the properties of a bootable device available in the
+// system.
+type BootOption struct {
+	common.Entity
+
+	// ODataContext is the odata context.
+	ODataContext string `json:"@odata.context"`
+	// ODataType is the odata type.
+	ODataType string `json:"@odata.type"`
+	// Alias is the alias of this boot source if one exists.
+	Alias BootSourceOverrideTarget
+	// BootOptionEnabled is an indication of whether the boot option is
+	// enabled. If true , it is enabled. If false, the boot option that the
+	// boot order array on the computer system contains is skipped. In the
+	// UEFI context, this property shall influence the load option active
+	// flag for the boot option.
+	BootOptionEnabled bool
+	// BootOptionReference is the unique identifier seen in Boot.BootOrder.
+	BootOptionReference string
+	// DisplayName is the user-readable display name of the boot option
+	// that appears in the boot order list in the user interface.
+	DisplayName string
+	// UefiDevicePath is the UEFI device path to access this UEFI boot
+	// option.
+	UefiDevicePath string
+}
+
+// GetBootOption will get a BootOption instance from the service.
+func GetBootOption(c common.Client, uri string) (*BootOption, error) {
+	resp, err := c.Get(uri)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	var bootoption BootOption
+	err = json.NewDecoder(resp.Body).Decode(&bootoption)
+	if err != nil {
+		return nil, err
+	}
+
+	bootoption.SetClient(c)
+	return &bootoption, nil
+}
+
 // ResetType describe the type off reset to be issue by the resource
 type ResetType string
